Updated Coin SFX
Updated Coin SFX so that it plays the full sound. By moving the SFX into the Main folder - I tied it to the player and referenced it from there to play whenever a coin was picked up.

Previously it was tied to the coin_3.tscn which was programmed to delete after an object entered it. Hence the full sound would never play.